requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
Flask==3.0.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
openai==0.28.1
platformdirs==3.11.0
pylint==3.0.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.0
telebot==0.0.5
tomlkit==0.12.1
tqdm==4.66.1
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
